backend/requirements.txt
attrs==24.2.0
beautifulsoup4==4.12.3
behave==1.2.6
blinker==1.8.2
cachelib==0.9.0
certifi==2024.8.30
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==44.0.0
cssselect==1.2.0
exceptiongroup==1.2.2
execnet==2.1.1
fasteners==0.19
filelock==3.15.4
Flask==3.0.3
Flask-Caching==2.3.0
Flask-Cors==5.0.0
Flask-SQLAlchemy==3.1.1
greenlet==3.0.3
gunicorn==23.0.0
h11==0.14.0
idna==3.8
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
outcome==1.3.0.post0
packaging==24.1
parameterized==0.9.0
parse==1.20.2
parse_type==0.6.3
pdbp==1.5.4
platformdirs==4.2.2
pluggy==1.5.0
psycopg2-binary==2.9.11
py==1.11.0
pycparser==2.22
Pygments==2.18.0
PyMySQL==1.1.1
pynose==1.5.2
pyotp==2.9.0
pyreadline3==3.4.1
PySocks==1.7.1
pytest==8.3.2
pytest-html==2.0.1
pytest-metadata==3.1.1
pytest-ordering==0.6
pytest-rerunfailures==14.0
pytest-xdist==3.6.1
python-dotenv==1.0.1
PyYAML==6.0.2
rapidfuzz==3.9.7
requests==2.31.0
rich==13.8.0
sbvirtualdisplay==1.3.0
selenium==4.24.0
seleniumbase==4.30.3
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
SQLAlchemy==2.0.34
tabcompleter==1.3.3
thefuzz==0.22.1
trio==0.26.2
trio-websocket==0.11.1
typing_extensions==4.12.2
urllib3==2.2.2
websocket-client==1.8.0
Werkzeug==3.0.4
wsproto==1.2.0
